user/serializers.py

from rest_framework import serializers
from user.models import CustomUser as User

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id','email', 'name', 'address', 'phone']
